Refactored
Hype Train has been refactored down from 3 events to a single event.

Its a bit messy at points but it does the trick
